fix type error

+++ datasets/dataset_generic.py
@@ -92,7 +92,7 @@ class Generic_WSI_Classification_Dataset(Dataset):
 		for p in patients:
 			locations = self.slide_data[self.slide_data['case_id'] == p].index.tolist()
 			assert len(locations) > 0
-			label = self.slide_data['label'][locations].values
+			label = self.slide_data['label'][locations].values.astype(int)
 			if patient_voting == 'max':
 				label = label.max() 
 			elif patient_voting == 'maj':
